+++ apps/api/src/common/middlewares/logger.middleware.ts
@@ -1,7 +1,6 @@
 import { Injectable, Logger, NestMiddleware } from "@nestjs/common";
 import { NextFunction, Request, Response } from "express";
 
-
 @Injectable()
 export class LoggerMiddleware implements NestMiddleware {
   private readonly logger = new Logger("LoggerMiddleware");