+++ apps/api/src/core/log/log.controller.ts
@@ -5,7 +5,6 @@ import { JwtAuthGuard } from "../../auth/guards/jwt-auth.guard";
 import { LogDto } from "./dto/log.dto";
 import { LogService } from "./log.service";
 
-
 @ApiTags("Log")
 @ApiBearerAuth()
 @Controller("project/:projectId/env/:envId/log")