+++ apps/api/src/core/log/log.module.ts
@@ -4,7 +4,6 @@ import { DatabaseModule } from "../../database";
 import { LogController } from "./log.controller";
 import { LogService } from "./log.service";
 
-
 @Module({
   imports: [DatabaseModule],
   controllers: [LogController],