+++ packages/browser-client/src/apis/index.ts
@@ -1,10 +1,10 @@
 /* tslint:disable */
 /* eslint-disable */
-export * from './AuthApi';
-export * from './ChatApi';
-export * from './EnvironmentApi';
-export * from './HealthApi';
-export * from './LogApi';
-export * from './MembershipApi';
-export * from './ProjectApi';
-export * from './TeamApi';
+export * from "./AuthApi";
+export * from "./ChatApi";
+export * from "./EnvironmentApi";
+export * from "./HealthApi";
+export * from "./LogApi";
+export * from "./MembershipApi";
+export * from "./ProjectApi";
+export * from "./TeamApi";