+++ apps/api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # api
 
+## 1.1.18
+
+### Patch Changes
+
+- patch
+- Updated dependencies
+  - @montelo/db@1.1.18
+
 ## 1.1.17
 
 ### Patch Changes