+++ apps/montelo/CHANGELOG.md
@@ -1,5 +1,16 @@
 # montelo
 
+## 1.1.18
+
+### Patch Changes
+
+- patch
+- Updated dependencies
+  - @montelo/cli@1.1.18
+  - @montelo/common@1.1.18
+  - @montelo/fizz@1.1.18
+  - @montelo/promptfile@1.1.18
+
 ## 1.1.17
 
 ### Patch Changes