+++ packages/fizz/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @montelo/fizz
 
+## 1.1.18
+
+### Patch Changes
+
+- patch
+- Updated dependencies
+  - @montelo/common@1.1.18
+
 ## 1.1.17
 
 ### Patch Changes